relocate findStock and findTag function to its seperate finle
