+++ database/Tag.js
@@ -12,16 +12,5 @@ const tagsSchema = new mongoose.Schema({
 
 const Tag = mongoose.model('tag', tagsSchema);
 
-const findTag = (params) => {
-  const splitParams = params.tag.split(',');
-  return Tag.find({ tag: { $in: splitParams } }, (err, data) => {
-    if (err) {
-      console.log('ERRR', err);
-    } else {
-      return data;
-    }
-  })};
-
 
 module.exports = Tag;
-module.exports = findTag;
